microgrid-ml-pipeline/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Data Processing
numpy==1.26.3
pandas==2.1.4
pyyaml==6.0.1

# Machine Learning (optional - for when real models are added)
# torch==2.1.2  # Uncomment when using real PyTorch models
# xgboost==2.0.3  # Uncomment when using real XGBoost models

# Optimization (optional)
# cvxpy==1.4.1  # For MPC optimization
# pyomo==6.7.0  # Alternative optimization framework

# Database (optional)
# psycopg2-binary==2.9.9  # PostgreSQL
# asyncpg==0.29.0  # Async PostgreSQL

# Caching (optional)
# redis==5.0.1

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
